+++ src/app/types/finalityProviders.ts
@@ -17,17 +17,8 @@ export interface Description {
   details: string;
 }
 
-export type FinalityProviderState = keyof typeof stateMap;
-
-const stateMap = {
-  FINALITY_PROVIDER_STATUS_INACTIVE: "Inactive",
-  FINALITY_PROVIDER_STATUS_ACTIVE: "Active",
-  FINALITY_PROVIDER_STATUS_JAILED: "Jailed",
-  FINALITY_PROVIDER_STATUS_SLASHED: "Slashed",
-};
-
-export const mapFinalityProviderState = (
-  state: FinalityProviderState,
-): string => {
-  return stateMap[state];
-};
+export type FinalityProviderState =
+  | "FINALITY_PROVIDER_STATUS_INACTIVE"
+  | "FINALITY_PROVIDER_STATUS_ACTIVE"
+  | "FINALITY_PROVIDER_STATUS_JAILED"
+  | "FINALITY_PROVIDER_STATUS_SLASHED";
